// src/app/models/User.ts
export class User {
    user_id: number;
    username: string;
    name: string;
    surname: string;
    indexNumber: string;
    jmbg: string;
    address: string;
    blocked:boolean;
  }

  export class User2 {
    user_id: number;
    username: string;
    name: string;
    password: string;
    surname: string;
    indexNumber: string;
    jmbg: string;
    address: string;
    blocked:boolean;
  }